Add multiple category search scenario spec.
The ability to search on multiple categories comes for free with
ActiveRecord and Postgres. We didn’t have to change the current query
to enable this, but I added the spec to make sure this functionality
exists and works.
